Database server connection

+++ src/main/java/gruppe_12_backend/rest_api_12/controller/StrainController.java
@@ -31,7 +31,7 @@ public class StrainController {
 
     @ApiOperation(value = "Get a strain by id", notes = "Returns a strain given the specific id")
     @GetMapping("/{id}")
-    public ResponseEntity<Strain> getStrain(@PathVariable("id") int id){
+    public ResponseEntity<Strain> getStrain(@PathVariable("id") long id){
         return ResponseEntity.ok(strainService.getStrain(id));
     }
