+++ src/main/java/gruppe_12_backend/rest_api_12/service/StrainService.java
@@ -17,7 +17,7 @@ public class StrainService {
         this.strainRepo = strainRepo;
     }
 
-    public Strain getStrain(int id){
+    public Strain getStrain(long id){
         return strainRepo.findById(id);
     }
